Feature test framework overhaul
There are a lot of features to cover.

The previous way led to all scenarios being tested
through a single method.  While it was convenient
enough to add scenarios, finding failed scenarios and
debugging them became burdensome.  VS+R#
and Rider both step through every scenario even
when a single scenario is selected for debug.

With this change, there are several updates.

- Each Scenarios class is identified as an independent test
  by xunit.  This makes it easier to find, run and debug
  failing scenarios.

- Documentation has been added to explain to help
  devs understand and write tests

- Scenario property names have been changed to make
  more sense for all use cases.  i.e. Help renamed to
  Result since Result can be used in non-help scenarios

- Models have been added to help ensure a sample of
  .net types are considered in applicable scenarios.
  The models also create a consistent pattern of use
  across scenarios.

- Several instances of ScenarioTestBase have been created
  as guides and moved to the FeatureTests folder for
  easy discovery