remove ICommand, IOption & IOperand interfaces
Removing interfaces in favor of concrete types will allow
us to add new members to the classes without creating
a breaking change.  This will be important to let us
continue to iterate on new features.

This does mean devs won't be able to extend commands
and arguments via inheritance.  It shouldn't be a problem
because they can extend using service classes in the
ContextData property.  We're already doing this for the
current middleware.