convert help providers to template pattern, clean dupe code
The DetailedHelpTextProvider and BasicHelpTextProvider
duplicated most of the same code.

If a dev wanted to override a section, they would need to
reimplement all of the logic.  This template pattern allows
a single method to be overridden, allowing devs to make
targetted changes to the output.

This also fixed some spacing inconsistencies which is
reflected in the tests in this PR