if an option is not specified, always throw an exception
ThrowOnUnexpectedArgument makes sense for operands but
not for options.

Ignoring unknown options has a problem of not knowing if
the next value was intended for that option or should
be the next argument in line.  This could lead to an option
value being processed as an argument leading to a bad state