make InterceptorInvocation publicly available
Changed the name of MiddlewareMethod to Interceptor.
I think this better describes how it's thought of and
reduces confusion with other middleware.

IInvocation.Invoke added a 'next' param which
will be null for the main command but populated
for interceptors

Discovered a possible bug where if someone
reassigned the invocations on the InvocationContext
they wouldn't have been honored because
we had to get the invocation from the ICommandDef.