Honor/call all interceptor methods in hierarchy
When there are several layers of nested commands,
it's unexpected behavior if options supplied for the top
command are no longer used because it's no longer
the immediate parent.  This violates the principal of
least surprise.

With this change, all interceptor methods will continue
to execute and inherited options can be specified at
the executed method, regardless of depth.

This also simplifies access to the invocation contexts
because they can be accessed as a collection. Several
middleware components were able to remove the
dependency on ICommandDef checks.

Populate SubCommand properties for access
in Interceptor methods