remove Command.HasInterceptor - purpose unclear
On review, I realized it was not clear if this would be
true if any parent had an interceptor or if this would
be true for a parent with a default command and there
isn't a need for it yet.  The need may eventually be
to know if there are interceptor options that can be
specified, but that also isn't a need yet.

Defering the feature