cleanup ArgumentTemplate and name resolution
Simplified how option names are provided.
Simplified how HelpOptionName is referenced.

Only one class needs to understand ArgumentTemplate now.
Only one location needs to parse a string to a short name.

It wasn't obvious from the code, but it turns out nothing
was setting TypeDisplayName from the template.  The code
to handle two sources for that was complicated and this
change is able to remove it.