introduce NameTransformation and CommandNodeType
NameTransformation enables other packages to modify names,
making it possible to extract Humanizer into a separate
package.

CommandNodeType captures the concept that commands
can contain 3 node types, commands, operands and options
and that some of those are called arguments.  I've been
trying to determine a name for this for a while and this is
the best so far.