enhanced prompting and password masking
- introduce Password type
- introduce IPrompter type
- prompt on missing options and operands
- included interceptor options
- exclude flags
- mask input for arguments of type Password
- optional filter for arguments so devs can filter on
  attributes or requiredness or ...
- test tools for testing prompts
- ctrl+c exits app during prompting
- escape can exit a prompt
- multi-value args add a value per line
- custom prompt and responder implementations
  can be used

A lot of work was done to understand how values are mapped
to ConsoleKeyInfo.  More work will need to be done to
support arrow keys, insert etc.  I've decided that's for
another day.  This should be adequate for now and is
certainly more robust that what we've had.

Supporting arrow keys will require exposing more methods on
IConsole to support setting position and determining if we
want to support moving up and down lines