Command.Options does not include inherited options
Inherited options should only be used for subcommands

- Move logic from mapping extensions to Command to ensure
inherited options are added to subcommands

- obsolete Inherited for AssignToExecutableSubcommands
as a more descriptive name for Option and
OptionAttribute

- update help, removing the clunky footnote concept.
since there will only be confusion when a default
command also has interceptor options, adding a new
options section is clear enough.

- leaving the term Inherited in the tests and comments
as it is more terse than AssignToExecutableSubcommands
and added Inherited term to the xml docs to make
the connection easier to see.