remove NormalizeLineEndings, stop trailing whitespc in help
This stops the test framework from dealing with
whitespace. It originally had to because the help providers
at the time output a lot of trailing whitespaces.

With the updated provider, we can trim the space per row.
After that, it was a couple small tweaks in ParseReporter
and now our tests exactly represent our output including
the whitespaces.