add IEnvironment and SystemEnvironment impl
to make it possible to test what is used by CommandDotNet
and for easier testing of other properties.

Added many, but not all properties.

Did not add NewLine because CommandDotNet uses
NewLine in a lot of places that IEnvironment couldn't be
injected to easily.  Adding NewLine could lead to surprise
when the framework doesn't use the value from
IEnvironment.

We may make Environment available via
and ambient property, but we'll need to consider more
carefully as it introduces a perf hit and an opportunity
for bugs